// clients/web/src/features/layout/composables/controller/appLayoutControllerTypes.ts
export type CreateChannelModalSubmitPayload =
  | {
      kind: 'channel';
      name: string;
      type: Exclude<EchoChannelType, 'selfRoles'>;
      categoryId: string;
      iconKey: string;
    }
  | { kind: 'category'; name: string };